# openapi.yaml
openapi: 3.1.0
info:
  title: Happy Path Marketing API
  version: "1.0.0"
  description: |
    This API allows access to Happy Path Marketing's core services:
    Follow-Up Pathway (CRM and automation), Ratings Boost Pro (review generation),
    and ServicePath (AI workflow and campaign tools). Designed for integration
    with AI agents via the Model Context Protocol (MCP).
  contact:
    email: admin@happypathmarketing.com
    url: https://www.happypathmarketing.com
    name: Happy Path Marketing
  termsOfService: https://happypath.marketing/hpm-terms-conditions
  license:
    name: Privacy Policy
    url: https://happypath.marketing/hpm-privacy-policy

servers:
  - url: https://happy-path-crm-api.onrender.com

tags:
  - name: Follow-Up Pathway
    description: Lead management and CRM workflows
  - name: Ratings Boost Pro
    description: Review and reputation automation
  - name: ServicePath
    description: AI workflow automations and client outreach tools
  - name: Food Tracker
    description: Voice meal logging, photo confirmation, nutrition coaching, allergy safety, and sale alerts

paths:
  /leads:
    post:
      operationId: submitLead
      tags: [Follow-Up Pathway]
      summary: Add a new lead to the CRM via your backend
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Lead'
      responses:
        "201":
          description: Lead successfully submitted

  /ratings/requests:
    post:
      operationId: sendReviewRequest
      tags: [Ratings Boost Pro]
      summary: Send a review request to customers
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewRequest'
      responses:
        "200":
          description: Review requests sent

  /ratings/summary:
    get:
      operationId: getReviewSummary
      tags: [Ratings Boost Pro]
      summary: Get a summary of customer reviews
      responses:
        "200":
          description: Review summary returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewSummary'

  /servicepath/automations:
    post:
      operationId: triggerAutomation
      tags: [ServicePath]
      summary: Trigger an AI automation workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowRequest'
      responses:
        "200":
          description: Workflow successfully triggered

  /servicepath/campaigns:
    post:
      operationId: launchCampaign
      tags: [ServicePath]
      summary: Launch an email or SMS campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignRequest'
      responses:
        "200":
          description: Campaign launched successfully

  /health:
    get:
      operationId: getHealth
      tags: [Food Tracker]
      summary: Check service health
      responses:
        "200":
          description: Service is healthy

  /food/users:
    post:
      operationId: createFoodUser
      tags: [Food Tracker]
      summary: Create a food tracking profile with goals, allergies, stores, and notification settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FoodUserCreate'
      responses:
        "201":
          description: Food user created

  /food/users/{user_id}:
    get:
      operationId: getFoodUser
      tags: [Food Tracker]
      summary: Get a food tracking profile
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        "200":
          description: Food user returned
        "404":
          description: User not found

  /food/entries:
    post:
      operationId: createFoodEntry
      tags: [Food Tracker]
      summary: Log a spoken meal and request photo confirmation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FoodEntryCreate'
      responses:
        "201":
          description: Food entry created and photo prompt returned

  /food/entries/{entry_id}/photo-confirmation:
    post:
      operationId: confirmFoodEntryPhoto
      tags: [Food Tracker]
      summary: Confirm recognized foods from a meal photo
      parameters:
        - $ref: '#/components/parameters/EntryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PhotoConfirmation'
      responses:
        "200":
          description: Food entry updated with confirmed foods

  /food/users/{user_id}/entries:
    get:
      operationId: listFoodEntries
      tags: [Food Tracker]
      summary: List a user's meal history
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        "200":
          description: Food entries returned

  /food/users/{user_id}/favorites:
    get:
      operationId: listFavoriteFoods
      tags: [Food Tracker]
      summary: List favorite foods used for personalization and sale alerts
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        "200":
          description: Favorite foods returned
    post:
      operationId: createFavoriteFood
      tags: [Food Tracker]
      summary: Save a favorite food and optional store preferences
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FavoriteFoodCreate'
      responses:
        "201":
          description: Favorite food saved

  /food/sales:
    post:
      operationId: createFoodSale
      tags: [Food Tracker]
      summary: Add a grocery or restaurant sale for matching against favorites
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FoodSaleCreate'
      responses:
        "201":
          description: Sale created

  /food/users/{user_id}/sale-alerts:
    get:
      operationId: getFoodSaleAlerts
      tags: [Food Tracker]
      summary: Find current sales that match favorite foods and stores
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        "200":
          description: Matching sale alerts returned

  /food/users/{user_id}/meal-suggestions:
    get:
      operationId: getMealSuggestions
      tags: [Food Tracker]
      summary: Suggest future meals based on goals, allergies, and liked foods
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        "200":
          description: Personalized meal suggestions returned

  /food/users/{user_id}/coach:
    get:
      operationId: getFoodCoach
      tags: [Food Tracker]
      summary: Get lean-muscle coaching and allergy-aware recommendations
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        "200":
          description: Coaching summary returned

components:
  parameters:
    UserId:
      name: user_id
      in: path
      required: true
      schema:
        type: string
    EntryId:
      name: entry_id
      in: path
      required: true
      schema:
        type: string
  schemas:
    Lead:
      type: object
      required: [name, email]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        source:
          type: string

    ReviewRequest:
      type: object
      required: [contacts]
      properties:
        contacts:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              email:
                type: string
              phone:
                type: string

    ReviewSummary:
      type: object
      properties:
        average_rating:
          type: number
        total_reviews:
          type: integer
        recent_comments:
          type: array
          items:
            type: string

    WorkflowRequest:
      type: object
      required: [workflow_id, lead_id]
      properties:
        workflow_id:
          type: string
        lead_id:
          type: string
        context:
          type: object
          additionalProperties: true

    CampaignRequest:
      type: object
      required: [type, message, targets]
      properties:
        type:
          type: string
          enum: [email, sms]
        message:
          type: string
        targets:
          type: array
          items:
            type: string


    FoodUserCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        goals:
          type: object
          additionalProperties: true
          example:
            primary: gain lean muscle
            protein_grams_per_day: 160
        allergies:
          type: array
          items:
            type: string
        favorite_stores:
          type: array
          items:
            type: string
        notification_preferences:
          type: object
          additionalProperties: true

    FoodEntryCreate:
      type: object
      required: [user_id, spoken_text]
      properties:
        user_id:
          type: string
        spoken_text:
          type: string
          description: Voice transcription such as "I ate chicken, rice, and broccoli."
        parsed_foods:
          type: array
          items:
            type: string
        photo_url:
          type: string
          format: uri
        confirmed_foods:
          type: array
          items:
            type: string
        liked:
          type: boolean
        meal_time:
          type: string
          format: date-time
        location:
          type: string
        estimated_nutrition:
          type: object
          additionalProperties: true
        notes:
          type: string

    PhotoConfirmation:
      type: object
      required: [confirmed_foods]
      properties:
        photo_url:
          type: string
          format: uri
        confirmed_foods:
          type: array
          items:
            type: string
        liked:
          type: boolean
        estimated_nutrition:
          type: object
          additionalProperties: true
        notes:
          type: string

    FavoriteFoodCreate:
      type: object
      required: [food_name]
      properties:
        food_name:
          type: string
        store_preferences:
          type: array
          items:
            type: string

    FoodSaleCreate:
      type: object
      required: [food_name, store_name]
      properties:
        food_name:
          type: string
        store_name:
          type: string
        store_location:
          type: string
        price:
          type: string
        sale_ends_at:
          type: string
          format: date-time
        source_url:
          type: string
          format: uri